+++ eds_scikit/plot/data_quality.py
@@ -1,15 +1,15 @@
 from copy import copy
 from datetime import datetime
-from typing import Optional, Tuple, Union
+from typing import Tuple
 
 import altair as alt
 import numpy as np
 import pandas as pd
-from pandas.core.frame import DataFrame
 from pandas.core.series import Series
 
 from ..utils.checks import check_columns
 from ..utils.framework import bd
+from ..utils.typing import DataFrame
 
 
 def plot_age_pyramid(
@@ -17,7 +17,7 @@ def plot_age_pyramid(
     datetime_ref: datetime = None,
     filename: str = None,
     savefig: bool = False,
-    return_vector: bool = False,
+    return_array: bool = False,
 ) -> Tuple[alt.Chart, Series]:
     """Plot an age pyramid from a 'person' pandas DataFrame.
 
@@ -57,15 +57,18 @@ def plot_age_pyramid(
             raise ValueError("You have to set a filename")
         if not isinstance(filename, str):
             raise ValueError(f"'filename' type must be str, got {type(filename)}")
+
     datetime_ref_raw = copy(datetime_ref)
-    person_ = person.copy()
+
     if datetime_ref is None:
         datetime_ref = datetime.today()
     elif isinstance(datetime_ref, datetime):
         datetime_ref = pd.to_datetime(datetime_ref)
     elif isinstance(datetime_ref, str):
-        if datetime_ref in person_.columns:
-            datetime_ref = person_[datetime_ref]
+        # A string type for datetime_ref could be either
+        # a column name or a datetime in string format.
+        if datetime_ref in person.columns:
+            datetime_ref = person[datetime_ref]
         else:
             datetime_ref = pd.to_datetime(
                 datetime_ref, errors="coerce"
@@ -80,25 +83,24 @@ def plot_age_pyramid(
             f"`datetime_ref` must be either None, a parseable string date"
             f", a column name or a datetime. Got type: {type(datetime_ref)}, {datetime_ref}"
         )
+
+    cols_to_keep = ["person_id", "birth_datetime", "gender_source_value"]
+    person_ = bd.to_pandas(person[cols_to_keep])
+
     person_["age"] = (datetime_ref - person_["birth_datetime"]).dt.total_seconds()
     person_["age"] /= 365 * 24 * 3600
 
+    mask = person_["age"] > 90
+    person_.loc[mask, "age"] = 99
+
     bins = np.arange(0, 100, 10)
     labels = [f"{left}-{right}" for left, right in zip(bins[:-1], bins[1:])]
-    person_["age_bins"] = bd.cut(person_["age"], bins=bins, labels=labels)
-
-    person_["age_bins"] = (
-        person_["age_bins"].astype(str).str.lower().str.replace("nan", "90+")
-    )
+    person_["age_bins"] = pd.cut(person_["age"], bins=bins, labels=labels)
 
     person_ = person_.loc[person_["gender_source_value"].isin(["m", "f"])]
     group_gender_age = person_.groupby(["gender_source_value", "age_bins"])[
         "person_id"
     ].count()
-
-    # Convert to pandas to ease plotting.
-    # Since we have aggregated the data, this operation won't crash.
-    group_gender_age = bd.to_pandas(group_gender_age)
 
     male = group_gender_age["m"].reset_index()
     female = group_gender_age["f"].reset_index()
@@ -136,10 +138,10 @@ def plot_age_pyramid(
 
     if savefig:
         chart.save(filename)
-        if return_vector:
+        if return_array:
             return group_gender_age
 
-    if return_vector:
+    if return_array:
         return chart, group_gender_age
 
     return chart