+++ changelog.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Adding person-dependant `datetime_ref` to `plot_age_pyramid`
+- Adding person-dependent `datetime_ref` to `plot_age_pyramid`
 
 ## v0.1.2 (2022-12-05)